CHG: register categories
we’re going to create one button, “Show me more…”, that will cause the app to launch when tapped. We’re also going to set the delegate property of the user notification center to be self, meaning that any alert-based messages that get sent will be routed to our view controller to be handled.